Pulling WeasylLookup class out to its own file, and renaming the investigation method
